+++ tests/TestCase.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace Pepperfm\Flashboard\Tests;
 
 use Illuminate\Config\Repository;
-use Illuminate\Container\Container;
 use Illuminate\Events\Dispatcher;
 use Illuminate\Routing\Router;
 use Illuminate\Support\Facades\Facade;
 use PHPUnit\Framework\TestCase as BaseTestCase;
+use Psr\Log\NullLogger;
 
 abstract class TestCase extends BaseTestCase
 {
-    protected Container $app;
+    protected TestApplication $app;
 
     protected Router $router;
 
@@ -22,8 +22,8 @@ abstract class TestCase extends BaseTestCase
         parent::setUp();
 
         Facade::clearResolvedInstances();
-        $this->app = new Container();
-        Container::setInstance($this->app);
+        $this->app = new TestApplication();
+        \Illuminate\Container\Container::setInstance($this->app);
         /** @phpstan-ignore-next-line */
         Facade::setFacadeApplication($this->app);
 
@@ -44,6 +44,7 @@ abstract class TestCase extends BaseTestCase
 
         $this->app->instance('config', $config);
         $this->app->instance('events', new Dispatcher($this->app));
+        $this->app->instance('log', new NullLogger());
         $this->router = new Router($this->app['events'], $this->app);
         $this->app->instance('router', $this->router);
         $this->app->alias('router', Router::class);